Laser Function
This is working and the colors change based on the key pressed (A, W, S, D). I'm couldn't get the space bar to work as well as the color change but this seemed to work out fine. If this will be an issue with the actual game play then I can go through and try to figure out another way to go about it.