Attributions and License Links
I added these both in separation (the above two) and together (the bottom one). Not sure which will work best but these are directly from the website freesound.org and in the format which the website suggests under the topic "How do I credit/attribute?".  I wasn't 100% on our game name so feel free to correct that should it be wrong. See you all tomorrow