Sound and Laser python files
I'm still trying to figure out the best way to work these into the game. The sound I don't think will be too difficult as I figure we can reference it with the command (cmd) as needed throughout the game. The laser piece might be more difficult as I was just changing the color based on certain keys being pressed. I'm not sure how that would work with the switch we have in game, also unsure how to go about derendering the line drawn for the laser anytime the key is not being pressed. Will continue working on it but wanted to get the files up here