Added sound to game
I added sound in the game and it does work. ((YAY)) Also I tried to get the laser lines to stop when they hit a enemy object but I have not been successful at this. I also changed the key recognized during the menu to the RETURN key from the SPACE key as the SPACE key was causing the firing sound to go off at the start of game prior to actually firing.